requirement.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2

opencv-python==4.10.0.84
pillow==10.4.0

transformers==4.44.2

pypdf==4.3.1
python-docx==1.1.2
beautifulsoup4==4.12.3
lxml==5.2.2

pandas==2.2.2
numpy==1.26.4
